update permissions; add mixins; update views; update tests; add setup_tests; update pre-commit
