+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
         stages: [push, commit]
 
     -   id: flake8
-        args: ['--ignore=F401,F811']
+        args: ['--ignore=F401,F811,F403']
         name: Run flake8
         entry: python -m flake8
         language: system
