+++ app/account/views.py
@@ -5,19 +5,17 @@ from flask_pydantic import validate
 
 from account.permissions import account_permission
 from account.schemas import RegisterUser, AccessToken, RefreshToken, AccountData
-from account.serializers import account_serializer
 from account.services_views import (
     create_new_account,
     create_tokens,
     generate_access_token_from_refresh,
-    authenticate,
-    get_user_from_pk,
-    update_account,
-    remove_account,
-    get_account
+    authenticate
 )
 from account.validate_datas import validate_register_account_data, validate_access_token_data
-from for_delete import func_rud_mixin
+
+from utils.data_process import validate_data_for_create_or_update
+from utils.endpoint_mixins import detail_endpoint_mixin, list_endpoint_mixin
+
 from models import Account
 
 auth_urls = Blueprint('auth', __name__, url_prefix='/auth')
@@ -70,41 +68,23 @@ def update_access_token(body: RefreshToken) -> tp.Tuple[Response, int]:
 @authenticate
 def user_detail(pk: int) -> tp.Tuple[Response, int]:
     """
-    Get user
+    Detail user
     """
 
-    return func_rud_mixin(schema=AccountData,
-                          model=Account,
-                          request=request,
-                          pk=pk,
-                          serializer=account_serializer)
-    # user_data, status_code = get_user_from_pk(pk=pk)
-    # if status_code == 404:
-    #     return jsonify(user_data), status_code
-    #
-    # has_permission = account_permission(user=request.user, request=request, pk=pk)  # type: ignore
-    #
-    # if not has_permission:
-    #     return jsonify({'Error': 'You has not permissions to perform this action'}), 401
-    #
-    # if request.method == 'PATCH':
-    #     data = AccountData(**request.json).__dict__  # type: ignore
-    #     for key in data.copy().keys():
-    #         if data[key] == '':
-    #             del data[key]
-    #     update_account(pk=pk, data=data)
-    # elif request.method == 'DELETE':
-    #     remove_account(pk=pk)
-    #     return jsonify({}), 204
-    # elif request.method == 'GET':
-    #     return jsonify(user_data), status_code
-    #
-    # user_data, status_code = get_user_from_pk(pk=pk)
-    # return jsonify(user_data), status_code
+    data = validate_data_for_create_or_update(schema=AccountData, request=request)
+    return detail_endpoint_mixin(schema=AccountData,
+                                 model=Account,
+                                 request=request,
+                                 pk=pk,
+                                 data=data,
+                                 permissions=[account_permission])
 
 
 @account_urls.route('/', methods=['GET'])
 @authenticate
-def get_users() -> tp.Tuple[Response, int]:
-    accounts, status_code = get_account()
-    return jsonify(accounts), status_code
+def list_users() -> tp.Tuple[Response, int]:
+    """
+    List user
+    """
+
+    return list_endpoint_mixin(schema=AccountData, model=Account, request=request)
