+++ app/models.py
@@ -18,19 +18,6 @@ class Account(db.Model):
     last_name = db.Column(db.String(length=100))
     is_staff = db.Column(db.BOOLEAN())
 
-    def __init__(self, username: str,
-                 password: str,
-                 email: str,
-                 first_name: str,
-                 last_name: str,
-                 is_staff: bool = False) -> None:
-        self.username = username
-        self.password = password
-        self.email = email
-        self.first_name = first_name
-        self.last_name = last_name
-        self.is_staff = is_staff
-
 
 class Video(db.Model):
     """
