+++ app/tests/test_account/test_validate_data.py
@@ -1,45 +1,7 @@
-import pytest
-
-import settings
-import typing as tp
-
-from models import Account
 from account.schemas import RegisterUser, AccessToken
 from account.validate_datas import validate_register_account_data, validate_access_token_data
 
-from main import register_flask_application, db
-
-from utils import make_password
-
-
-@pytest.fixture(scope='class', autouse=True)
-def setup() -> tp.Any:
-    app_1 = register_flask_application(settings.TESTING_CONFIGURATION)
-    app_1.app_context().push()
-    db.init_app(app_1)
-    db.create_all()
-
-    password = make_password(password='password')
-
-    account_1 = Account(username='username',
-                        password=password,
-                        is_staff=False,
-                        email='email_example@mail.ru',
-                        first_name='first_name',
-                        last_name='last_name')
-    account_2 = Account(username='username_2',
-                        password=password,
-                        is_staff=False,
-                        email='email_example2@mail.ru',
-                        first_name='first_name2',
-                        last_name='last_name2')
-
-    db.session.add(account_1)
-    db.session.add(account_2)
-    db.session.commit()
-    yield app_1
-    db.session.remove()
-    db.drop_all()
+from tests.setup_tests import *
 
 
 class TestValidateRegisterAccountData:
