+++ app/tests/test_account/test_views_auth.py
@@ -1,7 +1,7 @@
 import json
 import typing as tp
 
-from tests.test_account.test_validate_data import setup
+from tests.setup_tests import *
 
 
 def test_user_register(setup: tp.Any) -> None:
